Fix crossorigin script module test assertions
The module/crossorigin.html WPT previously expected different behavior
when a "crossorigin" attribute was present on the script element vs
when it was missing. But that is incorrect per-spec. Updated the test
to expect the script load to be blocked in both cases.

BUG=594639

Review-Url: https://codereview.chromium.org/2914273002
Cr-Original-Commit-Position: refs/heads/master@{#476655}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 52647ef2089ff4554a6245d2a8ed9dc8699a5f44