[DevTools] Remove createRawLocationByURL call from createRawLocation
This patch stops calling createRawLocationByURL from createRawLocation.
If client passes a valid Script object, it's reasonable that created location
belongs to that exact script, and not some other script.

The only valid user of old behavior was ResourceScriptMapping, which
wants to find a suitable script inside of an html resource. It does it
explicitly now.

BUG=none

Review-Url: https://codereview.chromium.org/2882323002
Cr-Original-Commit-Position: refs/heads/master@{#476757}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c74ea9411865a9864c37a907e1d60f356969503b