Re-commit "win: Set is_clang=true by default" over the weekend.
This relands https://codereview.chromium.org/2870543003

To figure out unknown problems. If this inconveniences you in any way, please
do let us know (ideally on crbug.com) – finding annoying things are part of the
motivation for this change.

If this breaks anything, please revert with prejudice -- but file a bug for
us about what went wrong. We intend to revert this on Sunday afternoon if it's not
reverted earlier due to breaking something.

This also switches win_clang to use MSVC as compiler, to make sure things stay
compilable with both compilers even while this is in.

Known problems:
* build times without goma are slower (but you can do goma builds with debug
  info, which are faster)
* incremental linking of webcore.dll doesn't work, so if you work on blink
  on windows links will be slow. If this affects you, please star
  https://crbug.com/717207 and make yourself heard.

This doesn't affect what we ship, which is currently off an PGO builder.

BUG=709690, 82385

Review-Url: https://codereview.chromium.org/2916933002
Cr-Original-Commit-Position: refs/heads/master@{#476884}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8476c55fd44915447233200c063fcd426dfe9807