Remove references to rare_inherited_data_ in ComputedStyle.
To allow changes to where fields are stored, ComputedStyle code should
not refer directly to a group, as that code will break when we change
groups. This patch removes references to rare_inherited_data_ in
ComputedStyle, replacing with generated or handwritten getters.

This patch also edits a callsite of
ComputedStyle::SetTextOrientation to make it uniform with other
setters in ComputedStyle.

BUG=710938

Review-Url: https://codereview.chromium.org/2918983002
Cr-Original-Commit-Position: refs/heads/master@{#476943}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7f9edaaa091bddbff260234e49b6c2988720c5a6