Keep the wrapper of FetchEvent alive while waiting for the preload response.
This CL uses ActiveScriptWrappable and implements HasPendingActivity() to keep the wrapper of FetchEvent alive while waiting for the preload response.

FetchEvents are already kept alive by ServiceWorkerGlobalScopeProxy's |pending_preload_fetch_events_| until the preload response.
But the resolver function is not called in the current implementation if GC happens. (https://crbug.com/728013#c28)
That is because if the wrapper of FetchEvent are GCed, ScriptPromiseProperty can't call the resolver.

BUG=728013

Review-Url: https://codereview.chromium.org/2922863002
Cr-Original-Commit-Position: refs/heads/master@{#476977}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7f1ac9b18fed816431798435d68d8ee90668d392