Stop crashing when VisibleSelections in DOM and Flat trees are inconsistent
There are some cases where FrameSelection::ComputeVisibleSelectionInDOMTree()
and FrameSelection::ComputeVisibleSelectionInFlatTree() return inconsistent
results, with one being null and another being non-null. This is unexpected
and causes crashes in some cases.

This patch stops the crash in FrameSelection::SelectionHasFocus() due to this
inconsistency. It doesn't fix the root cause, though, as the inconsistency is
still there. It is made small to be merged to M60.

BUG=725457
TEST=FrameSelectionTest.InconsistentVisibleSelectionNoCrash

Review-Url: https://codereview.chromium.org/2920583003
Cr-Original-Commit-Position: refs/heads/master@{#477144}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0cb66cf7b8cbc857c590ca5499f0ad6828684702