Introduce abstract StyleableMarker subclass of DocumentMarker
This CL refactors some code in CompositionMarker into a new abstract class
StyleableMarker, which will contain shared logic for MarkerTypes that can be
rendered with different underline color/thickness and/or background color.

This is in preparation for adding two new MarkerTypes I need to support some
features on Android (one for marking the range of text an open spellcheck/
suggestion menu pertains to, the other for marking ranges of text the user can
open a suggestion menu for by tapping on them).

BUG=715365

Review-Url: https://codereview.chromium.org/2922233002
Cr-Original-Commit-Position: refs/heads/master@{#477221}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 44d1353d9f4b134e7aee2b1d70eae0e0320b42c7