Move AddProperty to a helper class.
This cl paves the way for parsing shorthand properties. AddProperty is
currently a private function within CSSPropertyParser. We now move it
to a helper class so that it can be accessed by property APIs when a
property parses its values and adds parsed properties to a set of all
parsed properties for further processing.

BUG=668012

Review-Url: https://codereview.chromium.org/2917323003
Cr-Original-Commit-Position: refs/heads/master@{#477261}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 97d2eb259484cb04b331ac05367ff2bb1c31ed5a