Add ActiveSuggestionMarker
This is a new type of DocumentMarker that will be used on Android to highlight
a region of text that an open spellcheck/suggestion menu pertains to. For a
spellcheck menu, we'll want to highlight the text in red (to match the
underline), but for a suggestion menu, we'll want to highlight the text in the
color of the suggestion underline (on Android, there are a handful of different
colors used based on which flags are set on the corresponding SuggestionSpan).
So I'm making this marker inherit from the StyleableMarker subclass I
refactored CompositionMarker on top of in
https://codereview.chromium.org/2922233002.

This CL doesn't actually add code to paint ActiveSuggestionMarkers; I will add
that in a future CL (together with layout tests).

BUG=715365

Review-Url: https://codereview.chromium.org/2923033002
Cr-Original-Commit-Position: refs/heads/master@{#477476}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cf561e1c071256c42e108ba8fd1fb71dcd1db57b