Handle ActiveSuggestionMarkers for accessibility
I'm adding a new type of DocumentMarker, ActiveSuggestionMarker, in
https://codereview.chromium.org/2923033002. This MarkerType will be used to mark
the region of text that an open spellcheck/suggestion menu pertains to. This
seems potentially useful to expose for accessibility, so I'm doing that in this
CL (I didn't want to add more complexity to the original CL).

BUG=715365

Review-Url: https://codereview.chromium.org/2931473002
Cr-Original-Commit-Position: refs/heads/master@{#477481}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b1d2f5d73ed7f1656733cc55c4abc8c716b47ee8