[Smart Text] Make SurroundingText work for input elements
The existing algorithm sets the maximal range of forward and backward
character iterators and then advances them. The maximal range used to
be the whole document, in this CL we limit the range to be inside the
root editable element if such element exists.

BUG=721840

Review-Url: https://codereview.chromium.org/2907963002
Cr-Original-Commit-Position: refs/heads/master@{#477572}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 48824066704f4c092f5193ae38f5c457d688813e