[Presentation API / Media Router] Relax PresentationRequest URL check.
The idea is that instead checking if the scheme of a Presentation URL is
supported in Blink, we will do that in MR. This patch defines a
whitelist of schemes supported by MR.

This patch also enforces the scheme check on the URL given during a
presentation availability call. PresentationServiceDelegateImpl will
report availability = false for URLs of unsupported schemes.

This also fixes a problem where you actually cannot use a cast:* URL
in a https side despite having been "whitelisted" in Blink, due to the
MixedContent check (which isn't needed, since the URL does not refer to
an actual network resource).

The integration tests for desktop are also updated with new URLs,
including test:*. See accompanying extension CL: cl/158049822

BUG=711541

Review-Url: https://codereview.chromium.org/2927503002
Cr-Original-Commit-Position: refs/heads/master@{#477735}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e86537c744f1410319ce3ea4659f8c022f67ece3