Add a feature for not compositing small scrollers
We find that more than ~60% scrollers(scrollbars) are smaller than
certain size and only 10% of them gets scrolled by users. As a result,
we are thinking of not compositing the small scrollers to save GPU
memory. The experiment is to investigate how much memory gains we could
get without compositing them. Having a finch experiment in place will
provide us with data for evaluating the GPU memory usage for not
compositing small scrollers.

This patch is to add a relative feature.

BUG=728267

Review-Url: https://codereview.chromium.org/2912403003
Cr-Original-Commit-Position: refs/heads/master@{#477753}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3abfb995607fad374795c7abd737c4d012e7d7b0