Add CompositorTouchAction to RuntimeEnabledFeatures.
We are now working on handling touch actions in cc, adding the
runtime enabled features would make it possible for the two
logics (main thread only and both main thread and cc) to co-exist.

The flag will also be useful when we lauch the feature in the
future.

BUG=727839

Review-Url: https://codereview.chromium.org/2930493002
Cr-Original-Commit-Position: refs/heads/master@{#478026}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4d6bb36e3fcfb9742b6428e43a8e5df12681d0d0