Remove DisplayItemClient keep-alive detection
It was added we we wanted to track short-lived DisplayItemClients
on canary builds. When a DisplayItemClient which has display items
in an non-invalidated cached subsequence is destructed, we could
get the exact stack trace.

The feature was broken when we refactored subsequence caching so
that it can no longer catch DisplayItemClient destruction in cached
subsequences. Instead of fixing it, this CL remove it.

just not at the accurate time when the short-lived DisplayItemClient
is destructed, but we can find the exact place by using other local
debugging techniques.

DisplayItemClient: :IsAlive() can still catch the same set of errors,
Change-Id: I9bb0c38bce974ae744b4dc2175589beca9b8822a
Reviewed-on: https://chromium-review.googlesource.com/528326
Reviewed-by: Chris harrelson <chrishtr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#478137}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe340067f4a02915935b27bcf45fcd860ca7caff