Streams: Fix exception message for NaN HWM
ReadableStream and WritableStream both used streamErrors.errInvalidHWM
instead of streamErrors.invalidHWM for the message when a strategy
highWaterMark was NaN. This resulted in a RangeError being thrown with
an empty message. Because the web platform tests don't look at the
message, they didn't catch the bug.

Fix the message. Also add a Chromium-specific regression test to
validate that it is correct.

BUG=719763

Review-Url: https://codereview.chromium.org/2933853003
Cr-Original-Commit-Position: refs/heads/master@{#478880}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7f4a499745ae9a8601e20ef616e8e1e877934269