Change DataPersistent constructor to allow one line initialisation.
Currently, DataPersistent objects must be initialised twice: once via
the constructor that sets it to nullptr and once more via the Init()
method that sets it to a nonnull value.

This patch changes the default constructor to do Init() instead and
adds an extra constructor that sets it to nullptr. This way, we can
initialise DataPersistent objects in one line (useful for constructor
initializer lists).

BUG=628043

Review-Url: https://codereview.chromium.org/2935893002
Cr-Original-Commit-Position: refs/heads/master@{#478883}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3adbcbaa43c1d5069da92d154a7382943c5a2881