Add blink::WebSocketHandshakeThrottle
Add a facility for WebSocket handshakes to be delayed or cancelled by the
embedder. By implementing blink::Platform::CreateWebSocketHandshakeThrottle()
the embedder can supply an implementation of blink::WebSocketHandshakeThrottle
to throttle WebSocket handshakes.

To avoid slowing down the normal case, the throttling runs in parallel with the
handshake. Only access to the new connection from Javascript is delayed.

The default behaviour is to have no throttling.

The implementation is in
blink::DocumentWebSocketChannel. DocumentWebSocketChannelTest has new tests for
the new functionality.

This is part of the implementation of Safe Browsing for WebSockets. See the
design doc at
https://docs.google.com/document/d/1iR3XMIQukqlXb6ajIHE91apHZAxyF_wvRoB5JGeJYPs/edit

BUG=644744

Review-Url: https://codereview.chromium.org/2930263002
Cr-Original-Commit-Position: refs/heads/master@{#479709}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bf9d1c2f8e5ea0366b4b086c7d0541bdc15ab176