Propagate the user gesture bit when requesting push messaging permission.
Currently, directly subscribing to push messages without having
previously been granted notifications permissions results in
a permission prompt to the user. This prompt is hardcoded to always tell
the permissions layer that it was made by a user gesture, which may or
may not be the case.

This CL plumbs the user gesture bit from the renderer to the browser
when PushManager.subscribe() is called. This means permission metrics
for push will now have the correct values recorded for the user gesture bit.

BUG=733483

Review-Url: https://codereview.chromium.org/2935333003
Cr-Original-Commit-Position: refs/heads/master@{#479965}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6ba3fa621ea462377f2d6f189ae1eb186f9cae16