Reland of "Add blink::WebSocketHandshakeThrottle"
The CL has been updated to add WebSocketHandshakeThrottle.h to
BUILD.gn. Otherwise it is identical to the original CL,
https://codereview.chromium.org/2930263002/, which was reverted.

  Original description:

Add a facility for WebSocket handshakes to be delayed or cancelled by the
embedder. By implementing blink::Platform::CreateWebSocketHandshakeThrottle()
the embedder can supply an implementation of blink::WebSocketHandshakeThrottle
to throttle WebSocket handshakes.

To avoid slowing down the normal case, the throttling runs in parallel with the
handshake. Only access to the new connection from Javascript is delayed.

The default behaviour is to have no throttling.

The implementation is in
blink::DocumentWebSocketChannel. DocumentWebSocketChannelTest has new tests for
the new functionality.

This is part of the implementation of Safe Browsing for WebSockets. See the
design doc at
https://docs.google.com/document/d/1iR3XMIQukqlXb6ajIHE91apHZAxyF_wvRoB5JGeJYPs/edit

BUG=644744
TBR=yhirano,haraken

Review-Url: https://codereview.chromium.org/2942173002
Cr-Original-Commit-Position: refs/heads/master@{#479991}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3df1a00a1b61c1663e1183f54cb2d6a0badb26ae