Change wpt-import to not accept a "wpt/css" arg.
Before the csswg-test merge, the import script took a wpt or css argument
to decide which repo to fetch. In the transition period, it was changed
to tolerate an argument but ignore it. This CL makes it give an error
and usage message if invoked with an extra unknown argument.

Change-Id: I8805a0bcb969f496476aadba54ecd9c8cda6c72e
Reviewed-on: https://chromium-review.googlesource.com/537983
Reviewed-by: Jeff Carpenter <jeffcarp@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#480092}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8aed8bc33387861bf09a42448f2b3efbbbc316e1