Add DocumentShutdown{Notifier,Observer}
This is a subclass of LifecycleNotifier that classes can use to be notified when
the Document they're attached to is shut down.
SynchronousMutation{Notifier,Observer} provides the same functionality but is
more performance sensitive because it also gets notified when
Node::appendChild() and Node::removeChild() are called.

The idea to add this came from the code review of this CL:
https://codereview.chromium.org/2931443003#msg72

BUG=715365

Review-Url: https://codereview.chromium.org/2939223003
Cr-Original-Commit-Position: refs/heads/master@{#480164}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: afd1411d6b0dbb13fc94314c745755fd76bd763b