In-place modification of drag related functionality present in LocalFrame
In-Place modification of drag functionality present in LocalFrame.
NodeImage() and DragImageForSelection() are irrelevant to LocalFrame,
and will be moved out from LocalFrame to drag releated files.
This CL does the in-place changes to the above functions to
make movement of code independent of C++ core changes.

BUG=727981

Review-Url: https://codereview.chromium.org/2934233002
Cr-Original-Commit-Position: refs/heads/master@{#480392}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c9cb177ab255f3978bcb203f0c0867d7eed1a80d