RTCPeerConnection map of tracks independent of streams added.
Instead of iterating local and remote streams to find tracks that
belong to a peer connection, the peer connection keeps a map of
tracks. This will allow tracks to be added to the peer connection that
do not have an associated stream, which is possible with the
to-be-implemented RTCPeerConnection.replaceTrack.

On local and remote streams being added their tracks are added to the
map. A MediaStreamObserver interface is added and implemented such
that if a track is added to one of the streams it will also be added
to the RTCPeerConnection's tracks_.

Using a WeakMember-based map, unused tracks will automatically be
removed when destroyed by the GC. Our own ref count is not necessary.

BUG=700916, 705901

Review-Url: https://codereview.chromium.org/2931413002
Cr-Original-Commit-Position: refs/heads/master@{#480502}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1d05904bcc684ee1267bcef56496a9c3ebce2b61