Worklet: Move ParseCredentialsOption() to anonymous namespace for cleanup
This is a follow-up CL for
https://chromium.googlesource.com/chromium/src/+/8f9e4bf115f530b797a615fdecd5711845fec39f

Worklet::ParseCredentialsOption() was called from both MainThreadWorklet and
ThreadedWorklet, but after the merge of them, the function is used only from
Worklet. Therefore, it's no longer necessary that the function is a static
member of Worklet.

Bug: 727194
Change-Id: Ic9a7c90e61ec1a6a619096500a2e30064199d09f
Reviewed-on: https://chromium-review.googlesource.com/540958
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#480749}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 17b0d447e8382efbc891bd5c519b484a325ec8c3