Fire start/stop based on Frame::IsLoading for same document navigations
Prior to this change, DidStartLoading/DidStopLoading is fired for
same document navigations based on a heuristic similar to (but not
exactly the same as) the heuristic that changes the Frame loading bit.
There's no reason to have multiple different definitions, and this
should prevent a situation in which we have interleaved start/stop
notifications (or 2 starts but only 1 stop).

BUG=

Review-Url: https://codereview.chromium.org/2943523002
Cr-Original-Commit-Position: refs/heads/master@{#480968}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f52c295d282cfb4c88f075008d53c9daec7da71e