Add LayoutTheme::PlatformActiveSpellingMarkerHighlightColor()
I'm working on adding support for the Android spellcheck menu in
https://codereview.chromium.org/2931443003

and we need to add another color to LayoutTheme to represent what color we
should highlight a misspelled word while the menu is active. Android uses the
same red color as the spellcheck underline but with some transparency added. We
decided that it's better to add a separate method to LayoutTheme for getting the
color than to assume that we'll want to do an alpha blend on every platform that
might ever support this menu.

BUG=715365

Review-Url: https://codereview.chromium.org/2953493002
Cr-Original-Commit-Position: refs/heads/master@{#481085}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c0e31689b5b37b8f128db8b8475fa7744e675d47