Move ObjectContentType entirely to HTMLPlugInElement
There's no need to go to LocalFrameClientImpl to determine how to load
a plugin element, that was a purely pre-blink necessity.

Simplify a bunch of logic around HTMLPlugInElement to assume that url_
and service_type_ are always set and correct by the time the
ObjectContentType is being determined.

BUG=

Review-Url: https://codereview.chromium.org/2949873005
Cr-Original-Commit-Position: refs/heads/master@{#481217}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f12fc43363084c86d32bf151fe94f2994a9df8c8