Don't copy testharness.js and idlharness.js on wpt import.
Background: In general we want to keep
LayoutTests/resources/testharness.js in sync with
LayoutTests/external/wpt/resources/testharness.js because we want
testharness tests inside of and outside of wpt to behave similarly.

However, the practice of keeping these files always in sync means that
it's possible for results of testharness tests outside of wpt to change
upon import. Currently, the wpt importer tries to download new baselines
for any tests that fail on import. This works most of the time but
sometimes means we download baselines that we didn't want to download.

Rationale for this change: To simplify the import process, I think it's
easier to make it so that imports can only affect tests in external/wpt.

I think there should be a separate process to sync the testharness.js
and idlharness.js files; this could be manual and occasional, or it
could be automatic. It could be a separate "mode" of the wpt-importer.

In this change:
 1. Simplify importer so it doesn't copy files over
 2. Simplify presubmit so it doesn't check that the files are in sync
 3. Change wpt-update-expectations so that it ignores all failures outside of external/wpt.


Bug: 685854
Change-Id: I8f39790313cd1429f9016418a76a0eecc3183016
Reviewed-on: https://chromium-review.googlesource.com/539818
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#481285}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5bbc2bf8150d69cc30463a9040de0e97f5caad28