Move WebUSB IDL for IDL test to /interfaces/webusb.idl
The new pattern for idlharness.js-based tests is to fetch the IDL
definitions from a file in this directory instead of including it inline
in the test. This should make it easier to update these definitions
automatically.

The untested IDL declarations are also rearranged for better readability.

Change-Id: Ic275fa195c4990802df8ad0a28e9eb0859cb93af
Reviewed-on: https://chromium-review.googlesource.com/543816
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#481322}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6983d53d73d053ed0712fe5202188d5b9bd43501