[device-ram] Update the <value> calculations to match the updated spec
After implmenting the original spec, we noticed that the returned value
did not provide a good representation of the actual device-ram.
The issue was discussed on the spec:
https://github.com/WICG/device-ram/issues/1 and the spec was updated.

The new <value> is a rounding (instead of flooring) of the actual
memory to a bound that can have only the 2 most significant bits set.
More details in the spec description at:
https://github.com/WICG/device-ram

Bug: chromium:718622
Change-Id: I33df6b0480b8e65a24fe4765839fb22633dfcd43
Reviewed-on: https://chromium-review.googlesource.com/546997
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Fadi Meawad <fmeawad@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#482306}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1d35dcc9ba194c0b1bbbc01bce7e04caed7022b0