Layout tests: Default to "out" everywhere for the build directory.
This would make it so that instead of defaulting to "xcodebuild" on Mac
or "build" on Windows, we default to "out" everywhere. This doesn't
remove the --build-directory option, which can be used to override
the build directory.

Change-Id: I68e638371b201742202f417166c89ab925df46ba
Reviewed-on: https://chromium-review.googlesource.com/546938
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#482278}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe8dd8c9b58740fd0d876fa2b7ce6ada78165c5d