Fix error throwing logic in registerPaint
Right now in the implementation of registerPaint, we check whether the
paintDefinitionMap[name] exists or not, before checking whether |name|
is an empty string or not. The order should be reversed and this CL does
the correction. After this, our implementation matches the spec.

Bug: 736800
Change-Id: I7731351520003dc1ba043feedd15d58de2ee1e1b
Reviewed-on: https://chromium-review.googlesource.com/548817
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#482403}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9beaa97d0c8c4d1ab60cd288f81bb1af77eb5505