[LayoutNG] Add grapheme cluster iterations to LazyLineBreakIterator
This patch adds grapheme cluster iterations to LazyLineBreakIterator
as a new LineBreakType.

The implementation uses NonSharedCharacterBreakIterator, but expose as
a LineBreakType in LazyLineBreakIterator. This addition makes
NGLineBreaker to implement 'break-word' easier.

This may also be used for a break type being discussed in issue 720205
in future.

BUG=636993, 720205

Review-Url: https://codereview.chromium.org/2960563002
Cr-Original-Commit-Position: refs/heads/master@{#482378}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eff4014db12a2e52fd22a9eb100d7ebadd98c6e0