Fix nullptr deref in InvalidateKeyframeEffect
Animation::InvalidateKeyframeEffect assumed that if it had content then
it had a target element. After implementing the Animation constructor,
it was possible to have created an Animation object with a
KeyframeEffect with no target element. This patch checks whether the
target is nullptr before dereferencing.

BUG=734721

Review-Url: https://codereview.chromium.org/2961573002
Cr-Original-Commit-Position: refs/heads/master@{#482558}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e4424b5ca106bdb36ad962a9daadb86131576e98