Clarify SQlite threadsafety compile flags.
The Chrome build was setting THREADSAFE=1 on sqlite. But sqlite does not use
this flag (at least any more).

The flag sqlite expects is SQLITE_THREADSAFE. The default is the most
threadsafety which is then what we were getting. This patch makes that choice
explicit. There should be no change in behavior since 1 is the default. The
return value from sqlite3_threadsafe was verified to be the same (=1) both with
and without this patch.

Review-Url: https://codereview.chromium.org/2963633002
Cr-Original-Commit-Position: refs/heads/master@{#482837}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1be4c8f7053d29f3dbc0f7d21d341c04d9e63acc