Optimize DocumentMarkerController::MarkerAtPosition()
This method is used in some selection-related operations to look for spellcheck
markers. Now that we have DocumentMarkerList::MarkersIntersectingRange() (added
in https://codereview.chromium.org/2952953002) which uses binary search, we no
longer need to do a linear search over all the markers in the node to implement
this method.

BUG=707867

Review-Url: https://codereview.chromium.org/2949333002
Cr-Original-Commit-Position: refs/heads/master@{#482871}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 28fc48705db361c2fd93cd302c0d15c2a1ed8da4