Fix remove_webcore_debug_symbols to avoid constant building
The original gn implementation of remove_webcore_debug_symbols (in
crrev.com/1716523003) didn't generate PDBs for some of the webkit DLLs
in component builds, but that caused problems because ninja still
expects them to be outputs. This caused the build to never be clean (on
Windows) when this flag is used.

This change gets these DLLs to use minimal_symbols (on Windows builds
where symbol_level != 0) so that a PDB is always generated when
expected.

Doing test builds on my Windows machine of blink_core.dll I see the
following full-link times and chrome.dll.pdb sizes when I do a debug
component build with symbol_level = 2 when remove_webcore_debug_symbols
is set to the following configs:

        symbols: 71 s, 348 MB
minimal_symbols: 49 s,  82 MB
     no_symbols: 49 s, missing - builds never clean

This suggests that minimal_symbols is a reasonable choice (no increase
in build times) and a good fix for the not-clean builds.

This change does not alter the behavior on other platforms because
they don't have the not-clean issue.

In addition to avoiding constant building this also means that there
will be enough symbols for full call stacks.

BUG=647525

Review-Url: https://codereview.chromium.org/2948363004
Cr-Original-Commit-Position: refs/heads/master@{#482917}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b919d95a7f8be9ca291d6bbff942cfb42d08d5a1