Stop exporting ComputedStyle.
Exporting this class alone is respoonsible for 25% of the size of
blink_core.ilk, and most of it doesn't have to be exported.
Manually export the few (non-inline) functions that need this.
This CL reduces the size of blink_core.ilk from 2.59GB to 1.94GB.
(I'm at #475088 -- the win is likely smaller, but still substantial,
after https://chromium-review.googlesource.com/543155)

(One side effect is that the address of the inline functions might
end up different in different dlls, but that seems ok.)

BUG=560475

Review-Url: https://codereview.chromium.org/2963773002
Cr-Original-Commit-Position: refs/heads/master@{#483051}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 23768d32b212f3dbb00965c87a1b892d3ae6f634